Queries the collections endpoint

Queries the appropriate collections/ endpoint depending on the arguments passed to it.

Usage

collections(collID = NULL, authenticated = TRUE)

Arguments

collID

Collection ID

authenticated

whether to send authenticated API requests (defaults to TRUE)

Value

A list with information about collection(s) and the number of products available

Examples

collections(authenticated = FALSE)
#> [1] "Found  28652435  products."
#> $links
#> $links[[1]]
#> $links[[1]]$type
#> [1] "Link"
#> 
#> $links[[1]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3ACLM-IODC?format=json"
#> 
#> $links[[1]]$mediaType
#> [1] "application/json"
#> 
#> $links[[1]]$title
#> [1] "EO:EUM:DAT:MSG:CLM-IODC"
#> 
#> $links[[1]]$datasetTitle
#> [1] "Cloud Mask - MSG - Indian Ocean"
#> 
#> $links[[1]]$rel
#> [1] "collection/details"
#> 
#> $links[[1]]$numberOfProducts
#> [1] 252823
#> 
#> 
#> $links[[2]]
#> $links[[2]]$type
#> [1] "Link"
#> 
#> $links[[2]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0082?format=json"
#> 
#> $links[[2]]$mediaType
#> [1] "application/json"
#> 
#> $links[[2]]$title
#> [1] "EO:EUM:DAT:0082"
#> 
#> $links[[2]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 1 - MFG - 63 degree"
#> 
#> $links[[2]]$rel
#> [1] "collection/details"
#> 
#> $links[[2]]$numberOfProducts
#> [1] 147421
#> 
#> 
#> $links[[3]]
#> $links[[3]]$type
#> [1] "Link"
#> 
#> $links[[3]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0237?format=json"
#> 
#> $links[[3]]$mediaType
#> [1] "application/json"
#> 
#> $links[[3]]$title
#> [1] "EO:EUM:DAT:0237"
#> 
#> $links[[3]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B Low Resolution (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[3]]$rel
#> [1] "collection/details"
#> 
#> $links[[3]]$numberOfProducts
#> [1] 12644
#> 
#> 
#> $links[[4]]
#> $links[[4]]$type
#> [1] "Link"
#> 
#> $links[[4]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0241?format=json"
#> 
#> $links[[4]]$mediaType
#> [1] "application/json"
#> 
#> $links[[4]]$title
#> [1] "EO:EUM:DAT:0241"
#> 
#> $links[[4]]$datasetTitle
#> [1] "Climate-quality Advanced Microwave Radiometer Level 2 Products (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[4]]$rel
#> [1] "collection/details"
#> 
#> $links[[4]]$numberOfProducts
#> [1] 13065
#> 
#> 
#> $links[[5]]
#> $links[[5]]$type
#> [1] "Link"
#> 
#> $links[[5]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0274?format=json"
#> 
#> $links[[5]]$mediaType
#> [1] "application/json"
#> 
#> $links[[5]]$title
#> [1] "EO:EUM:DAT:0274"
#> 
#> $links[[5]]$datasetTitle
#> [1] "RINEX Auxiliary Data - Sentinel-6"
#> 
#> $links[[5]]$rel
#> [1] "collection/details"
#> 
#> $links[[5]]$numberOfProducts
#> [1] 13714
#> 
#> 
#> $links[[6]]
#> $links[[6]]$type
#> [1] "Link"
#> 
#> $links[[6]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0855?format=json"
#> 
#> $links[[6]]$mediaType
#> [1] "application/json"
#> 
#> $links[[6]]$title
#> [1] "EO:EUM:DAT:0855"
#> 
#> $links[[6]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 High Resolution - Sentinel-6"
#> 
#> $links[[6]]$rel
#> [1] "collection/details"
#> 
#> $links[[6]]$numberOfProducts
#> [1] 15500
#> 
#> 
#> $links[[7]]
#> $links[[7]]$type
#> [1] "Link"
#> 
#> $links[[7]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0857?format=json"
#> 
#> $links[[7]]$mediaType
#> [1] "application/json"
#> 
#> $links[[7]]$title
#> [1] "EO:EUM:DAT:0857"
#> 
#> $links[[7]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2P Low Resolution - Sentinel-6"
#> 
#> $links[[7]]$rel
#> [1] "collection/details"
#> 
#> $links[[7]]$numberOfProducts
#> [1] 11807
#> 
#> 
#> $links[[8]]
#> $links[[8]]$type
#> [1] "Link"
#> 
#> $links[[8]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0584?format=json"
#> 
#> $links[[8]]$mediaType
#> [1] "application/json"
#> 
#> $links[[8]]$title
#> [1] "EO:EUM:DAT:0584"
#> 
#> $links[[8]]$datasetTitle
#> [1] "SRAL Level 1B (version BC004) - Sentinel-3 - Reprocessed"
#> 
#> $links[[8]]$rel
#> [1] "collection/details"
#> 
#> $links[[8]]$numberOfProducts
#> [1] 57065
#> 
#> 
#> $links[[9]]
#> $links[[9]]$type
#> [1] "Link"
#> 
#> $links[[9]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0412?format=json"
#> 
#> $links[[9]]$mediaType
#> [1] "application/json"
#> 
#> $links[[9]]$title
#> [1] "EO:EUM:DAT:0412"
#> 
#> $links[[9]]$datasetTitle
#> [1] "SLSTR Level 2 Sea Surface Temperature (SST) - Sentinel-3"
#> 
#> $links[[9]]$rel
#> [1] "collection/details"
#> 
#> $links[[9]]$numberOfProducts
#> [1] 67864
#> 
#> 
#> $links[[10]]
#> $links[[10]]$type
#> [1] "Link"
#> 
#> $links[[10]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3AMSG15-RSS?format=json"
#> 
#> $links[[10]]$mediaType
#> [1] "application/json"
#> 
#> $links[[10]]$title
#> [1] "EO:EUM:DAT:MSG:MSG15-RSS"
#> 
#> $links[[10]]$datasetTitle
#> [1] "Rapid Scan High Rate SEVIRI Level 1.5 Image Data - MSG"
#> 
#> $links[[10]]$rel
#> [1] "collection/details"
#> 
#> $links[[10]]$numberOfProducts
#> [1] 1508372
#> 
#> 
#> $links[[11]]
#> $links[[11]]$type
#> [1] "Link"
#> 
#> $links[[11]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMULT%3AHIRSL1?format=json"
#> 
#> $links[[11]]$mediaType
#> [1] "application/json"
#> 
#> $links[[11]]$title
#> [1] "EO:EUM:DAT:MULT:HIRSL1"
#> 
#> $links[[11]]$datasetTitle
#> [1] "HIRS Level 1B - Metop - Global"
#> 
#> $links[[11]]$rel
#> [1] "collection/details"
#> 
#> $links[[11]]$numberOfProducts
#> [1] 134985
#> 
#> 
#> $links[[12]]
#> $links[[12]]$type
#> [1] "Link"
#> 
#> $links[[12]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0836?format=json"
#> 
#> $links[[12]]$mediaType
#> [1] "application/json"
#> 
#> $links[[12]]$title
#> [1] "EO:EUM:DAT:0836"
#> 
#> $links[[12]]$datasetTitle
#> [1] "SRAL Level 1A Unpacked L0 Complex Echoes (version BC005) - Sentinel-3 - Reprocessed"
#> 
#> $links[[12]]$rel
#> [1] "collection/details"
#> 
#> $links[[12]]$numberOfProducts
#> [1] 119939
#> 
#> 
#> $links[[13]]
#> $links[[13]]$type
#> [1] "Link"
#> 
#> $links[[13]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AASCSZF1B?format=json"
#> 
#> $links[[13]]$mediaType
#> [1] "application/json"
#> 
#> $links[[13]]$title
#> [1] "EO:EUM:DAT:METOP:ASCSZF1B"
#> 
#> $links[[13]]$datasetTitle
#> [1] "ASCAT Level 1 Sigma0 Full Resolution - Metop - Global"
#> 
#> $links[[13]]$rel
#> [1] "collection/details"
#> 
#> $links[[13]]$numberOfProducts
#> [1] 121798
#> 
#> 
#> $links[[14]]
#> $links[[14]]$type
#> [1] "Link"
#> 
#> $links[[14]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AASCSZR1B?format=json"
#> 
#> $links[[14]]$mediaType
#> [1] "application/json"
#> 
#> $links[[14]]$title
#> [1] "EO:EUM:DAT:METOP:ASCSZR1B"
#> 
#> $links[[14]]$datasetTitle
#> [1] "ASCAT Level 1 Sigma0 resampled at 12.5 km Swath Grid - Metop - Global"
#> 
#> $links[[14]]$rel
#> [1] "collection/details"
#> 
#> $links[[14]]$numberOfProducts
#> [1] 123155
#> 
#> 
#> $links[[15]]
#> $links[[15]]$type
#> [1] "Link"
#> 
#> $links[[15]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0394?format=json"
#> 
#> $links[[15]]$mediaType
#> [1] "application/json"
#> 
#> $links[[15]]$title
#> [1] "EO:EUM:DAT:0394"
#> 
#> $links[[15]]$datasetTitle
#> [1] "Land Surface Temperature with Directional Effects - MSG"
#> 
#> $links[[15]]$rel
#> [1] "collection/details"
#> 
#> $links[[15]]$numberOfProducts
#> [1] 19210
#> 
#> 
#> $links[[16]]
#> $links[[16]]$type
#> [1] "Link"
#> 
#> $links[[16]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0882?format=json"
#> 
#> $links[[16]]$mediaType
#> [1] "application/json"
#> 
#> $links[[16]]$title
#> [1] "EO:EUM:DAT:0882"
#> 
#> $links[[16]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 2 - MFG - 63 degree"
#> 
#> $links[[16]]$rel
#> [1] "collection/details"
#> 
#> $links[[16]]$numberOfProducts
#> [1] 143264
#> 
#> 
#> $links[[17]]
#> $links[[17]]$type
#> [1] "Link"
#> 
#> $links[[17]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0862?format=json"
#> 
#> $links[[17]]$mediaType
#> [1] "application/json"
#> 
#> $links[[17]]$title
#> [1] "EO:EUM:DAT:0862"
#> 
#> $links[[17]]$datasetTitle
#> [1] "AVHRR Fundamental Data Record - Release 1 - Multimission"
#> 
#> $links[[17]]$rel
#> [1] "collection/details"
#> 
#> $links[[17]]$numberOfProducts
#> [1] 752690
#> 
#> 
#> $links[[18]]
#> $links[[18]]$type
#> [1] "Link"
#> 
#> $links[[18]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0081?format=json"
#> 
#> $links[[18]]$mediaType
#> [1] "application/json"
#> 
#> $links[[18]]$title
#> [1] "EO:EUM:DAT:0081"
#> 
#> $links[[18]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 1 - MFG - 57 degree"
#> 
#> $links[[18]]$rel
#> [1] "collection/details"
#> 
#> $links[[18]]$numberOfProducts
#> [1] 176339
#> 
#> 
#> $links[[19]]
#> $links[[19]]$type
#> [1] "Link"
#> 
#> $links[[19]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0239?format=json"
#> 
#> $links[[19]]$mediaType
#> [1] "application/json"
#> 
#> $links[[19]]$title
#> [1] "EO:EUM:DAT:0239"
#> 
#> $links[[19]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 Low Resolution (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[19]]$rel
#> [1] "collection/details"
#> 
#> $links[[19]]$numberOfProducts
#> [1] 12644
#> 
#> 
#> $links[[20]]
#> $links[[20]]$type
#> [1] "Link"
#> 
#> $links[[20]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ACM%3AMETOP%3AASCSZFR02?format=json"
#> 
#> $links[[20]]$mediaType
#> [1] "application/json"
#> 
#> $links[[20]]$title
#> [1] "EO:EUM:CM:METOP:ASCSZFR02"
#> 
#> $links[[20]]$datasetTitle
#> [1] "ASCAT Level 1 SZF Climate Data Record Release 2 - Metop"
#> 
#> $links[[20]]$rel
#> [1] "collection/details"
#> 
#> $links[[20]]$numberOfProducts
#> [1] 37304
#> 
#> 
#> $links[[21]]
#> $links[[21]]$type
#> [1] "Link"
#> 
#> $links[[21]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0405?format=json"
#> 
#> $links[[21]]$mediaType
#> [1] "application/json"
#> 
#> $links[[21]]$title
#> [1] "EO:EUM:DAT:0405"
#> 
#> $links[[21]]$datasetTitle
#> [1] "Atmospheric Motion Vectors Climate Data Record Release 2 - MFG and MSG - 0 degree"
#> 
#> $links[[21]]$rel
#> [1] "collection/details"
#> 
#> $links[[21]]$numberOfProducts
#> [1] 437627
#> 
#> 
#> $links[[22]]
#> $links[[22]]$type
#> [1] "Link"
#> 
#> $links[[22]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0406?format=json"
#> 
#> $links[[22]]$mediaType
#> [1] "application/json"
#> 
#> $links[[22]]$title
#> [1] "EO:EUM:DAT:0406"
#> 
#> $links[[22]]$datasetTitle
#> [1] "SRAL Level 1B - Sentinel-3"
#> 
#> $links[[22]]$rel
#> [1] "collection/details"
#> 
#> $links[[22]]$numberOfProducts
#> [1] 34744
#> 
#> 
#> $links[[23]]
#> $links[[23]]$type
#> [1] "Link"
#> 
#> $links[[23]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0410?format=json"
#> 
#> $links[[23]]$mediaType
#> [1] "application/json"
#> 
#> $links[[23]]$title
#> [1] "EO:EUM:DAT:0410"
#> 
#> $links[[23]]$datasetTitle
#> [1] "OLCI Level 1B Reduced Resolution - Sentinel-3"
#> 
#> $links[[23]]$rel
#> [1] "collection/details"
#> 
#> $links[[23]]$numberOfProducts
#> [1] 35606
#> 
#> 
#> $links[[24]]
#> $links[[24]]$type
#> [1] "Link"
#> 
#> $links[[24]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0413?format=json"
#> 
#> $links[[24]]$mediaType
#> [1] "application/json"
#> 
#> $links[[24]]$title
#> [1] "EO:EUM:DAT:0413"
#> 
#> $links[[24]]$datasetTitle
#> [1] "SRAL Level 1A Unpacked L0 Complex echos - Sentinel-3"
#> 
#> $links[[24]]$rel
#> [1] "collection/details"
#> 
#> $links[[24]]$numberOfProducts
#> [1] 24156
#> 
#> 
#> $links[[25]]
#> $links[[25]]$type
#> [1] "Link"
#> 
#> $links[[25]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0415?format=json"
#> 
#> $links[[25]]$mediaType
#> [1] "application/json"
#> 
#> $links[[25]]$title
#> [1] "EO:EUM:DAT:0415"
#> 
#> $links[[25]]$datasetTitle
#> [1] "SRAL Level 2 Altimetry Global - Sentinel-3"
#> 
#> $links[[25]]$rel
#> [1] "collection/details"
#> 
#> $links[[25]]$numberOfProducts
#> [1] 34553
#> 
#> 
#> $links[[26]]
#> $links[[26]]$type
#> [1] "Link"
#> 
#> $links[[26]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0416?format=json"
#> 
#> $links[[26]]$mediaType
#> [1] "application/json"
#> 
#> $links[[26]]$title
#> [1] "EO:EUM:DAT:0416"
#> 
#> $links[[26]]$datasetTitle
#> [1] "SLSTR Level 2 Aerosol Optical Depth - Sentinel-3"
#> 
#> $links[[26]]$rel
#> [1] "collection/details"
#> 
#> $links[[26]]$numberOfProducts
#> [1] 685501
#> 
#> 
#> $links[[27]]
#> $links[[27]]$type
#> [1] "Link"
#> 
#> $links[[27]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0617?format=json"
#> 
#> $links[[27]]$mediaType
#> [1] "application/json"
#> 
#> $links[[27]]$title
#> [1] "EO:EUM:DAT:0617"
#> 
#> $links[[27]]$datasetTitle
#> [1] "Optimal Cloud Analysis Climate Data Record Release 1 - MSG - 0 degree"
#> 
#> $links[[27]]$rel
#> [1] "collection/details"
#> 
#> $links[[27]]$numberOfProducts
#> [1] 542908
#> 
#> 
#> $links[[28]]
#> $links[[28]]$type
#> [1] "Link"
#> 
#> $links[[28]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0301?format=json"
#> 
#> $links[[28]]$mediaType
#> [1] "application/json"
#> 
#> $links[[28]]$title
#> [1] "EO:EUM:DAT:0301"
#> 
#> $links[[28]]$datasetTitle
#> [1] "GSA Level 2 Climate Data Record Release 2 - MFG - 57 degree"
#> 
#> $links[[28]]$rel
#> [1] "collection/details"
#> 
#> $links[[28]]$numberOfProducts
#> [1] 378
#> 
#> 
#> $links[[29]]
#> $links[[29]]$type
#> [1] "Link"
#> 
#> $links[[29]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0851?format=json"
#> 
#> $links[[29]]$mediaType
#> [1] "application/json"
#> 
#> $links[[29]]$title
#> [1] "EO:EUM:DAT:0851"
#> 
#> $links[[29]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B High Resolution - Sentinel-6"
#> 
#> $links[[29]]$rel
#> [1] "collection/details"
#> 
#> $links[[29]]$numberOfProducts
#> [1] 15509
#> 
#> 
#> $links[[30]]
#> $links[[30]]$type
#> [1] "Link"
#> 
#> $links[[30]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0856?format=json"
#> 
#> $links[[30]]$mediaType
#> [1] "application/json"
#> 
#> $links[[30]]$title
#> [1] "EO:EUM:DAT:0856"
#> 
#> $links[[30]]$datasetTitle
#> [1] "Climate-quality Advanced Microwave Radiometer Level 2 Products - Sentinel-6"
#> 
#> $links[[30]]$rel
#> [1] "collection/details"
#> 
#> $links[[30]]$numberOfProducts
#> [1] 11359
#> 
#> 
#> $links[[31]]
#> $links[[31]]$type
#> [1] "Link"
#> 
#> $links[[31]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AASCSZO1B?format=json"
#> 
#> $links[[31]]$mediaType
#> [1] "application/json"
#> 
#> $links[[31]]$title
#> [1] "EO:EUM:DAT:METOP:ASCSZO1B"
#> 
#> $links[[31]]$datasetTitle
#> [1] "ASCAT Level 1 Sigma0 resampled at 25 km Swath Grid - Metop - Global"
#> 
#> $links[[31]]$rel
#> [1] "collection/details"
#> 
#> $links[[31]]$numberOfProducts
#> [1] 123154
#> 
#> 
#> $links[[32]]
#> $links[[32]]$type
#> [1] "Link"
#> 
#> $links[[32]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AGOMEL1?format=json"
#> 
#> $links[[32]]$mediaType
#> [1] "application/json"
#> 
#> $links[[32]]$title
#> [1] "EO:EUM:DAT:METOP:GOMEL1"
#> 
#> $links[[32]]$datasetTitle
#> [1] "GOME-2 Level 1B - Metop - Global"
#> 
#> $links[[32]]$rel
#> [1] "collection/details"
#> 
#> $links[[32]]$numberOfProducts
#> [1] 51113
#> 
#> 
#> $links[[33]]
#> $links[[33]]$type
#> [1] "Link"
#> 
#> $links[[33]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0833?format=json"
#> 
#> $links[[33]]$mediaType
#> [1] "application/json"
#> 
#> $links[[33]]$title
#> [1] "EO:EUM:DAT:0833"
#> 
#> $links[[33]]$datasetTitle
#> [1] "SRAL Level 1B (version BC005) - Sentinel-3 - Reprocessed"
#> 
#> $links[[33]]$rel
#> [1] "collection/details"
#> 
#> $links[[33]]$numberOfProducts
#> [1] 121571
#> 
#> 
#> $links[[34]]
#> $links[[34]]$type
#> [1] "Link"
#> 
#> $links[[34]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0835?format=json"
#> 
#> $links[[34]]$mediaType
#> [1] "application/json"
#> 
#> $links[[34]]$title
#> [1] "EO:EUM:DAT:0835"
#> 
#> $links[[34]]$datasetTitle
#> [1] "SRAL Level 1B Stack Echoes (version BC005) - Sentinel-3 - Reprocessed"
#> 
#> $links[[34]]$rel
#> [1] "collection/details"
#> 
#> $links[[34]]$numberOfProducts
#> [1] 119939
#> 
#> 
#> $links[[35]]
#> $links[[35]]$type
#> [1] "Link"
#> 
#> $links[[35]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0398?format=json"
#> 
#> $links[[35]]$mediaType
#> [1] "application/json"
#> 
#> $links[[35]]$title
#> [1] "EO:EUM:DAT:0398"
#> 
#> $links[[35]]$datasetTitle
#> [1] "Fire Risk Map - Released Energy Based - MSG"
#> 
#> $links[[35]]$rel
#> [1] "collection/details"
#> 
#> $links[[35]]$numberOfProducts
#> [1] 1003
#> 
#> 
#> $links[[36]]
#> $links[[36]]$type
#> [1] "Link"
#> 
#> $links[[36]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0880?format=json"
#> 
#> $links[[36]]$mediaType
#> [1] "application/json"
#> 
#> $links[[36]]$title
#> [1] "EO:EUM:DAT:0880"
#> 
#> $links[[36]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 2 - MFG - 0 degree"
#> 
#> $links[[36]]$rel
#> [1] "collection/details"
#> 
#> $links[[36]]$numberOfProducts
#> [1] 402755
#> 
#> 
#> $links[[37]]
#> $links[[37]]$type
#> [1] "Link"
#> 
#> $links[[37]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0647?format=json"
#> 
#> $links[[37]]$mediaType
#> [1] "application/json"
#> 
#> $links[[37]]$title
#> [1] "EO:EUM:DAT:0647"
#> 
#> $links[[37]]$datasetTitle
#> [1] "HIRS Level 1C Fundamental Data Record Release 1 - Multimission - Global"
#> 
#> $links[[37]]$rel
#> [1] "collection/details"
#> 
#> $links[[37]]$numberOfProducts
#> [1] 657201
#> 
#> 
#> $links[[38]]
#> $links[[38]]$type
#> [1] "Link"
#> 
#> $links[[38]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AGLB-SST-NC?format=json"
#> 
#> $links[[38]]$mediaType
#> [1] "application/json"
#> 
#> $links[[38]]$title
#> [1] "EO:EUM:DAT:METOP:GLB-SST-NC"
#> 
#> $links[[38]]$datasetTitle
#> [1] "Global L3C AVHRR Sea Surface Temperature (GHRSST) - Metop"
#> 
#> $links[[38]]$rel
#> [1] "collection/details"
#> 
#> $links[[38]]$numberOfProducts
#> [1] 5466
#> 
#> 
#> $links[[39]]
#> $links[[39]]$type
#> [1] "Link"
#> 
#> $links[[39]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3ACLM?format=json"
#> 
#> $links[[39]]$mediaType
#> [1] "application/json"
#> 
#> $links[[39]]$title
#> [1] "EO:EUM:DAT:MSG:CLM"
#> 
#> $links[[39]]$datasetTitle
#> [1] "Cloud Mask - MSG - 0 degree"
#> 
#> $links[[39]]$rel
#> [1] "collection/details"
#> 
#> $links[[39]]$numberOfProducts
#> [1] 736132
#> 
#> 
#> $links[[40]]
#> $links[[40]]$type
#> [1] "Link"
#> 
#> $links[[40]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0143?format=json"
#> 
#> $links[[40]]$mediaType
#> [1] "application/json"
#> 
#> $links[[40]]$title
#> [1] "EO:EUM:DAT:0143"
#> 
#> $links[[40]]$datasetTitle
#> [1] "Poseidon-4 Level 3 Wind/Wave Products Low Resolution in NRT - Sentinel-6"
#> 
#> $links[[40]]$rel
#> [1] "collection/details"
#> 
#> $links[[40]]$numberOfProducts
#> [1] 6029
#> 
#> 
#> $links[[41]]
#> $links[[41]]$type
#> [1] "Link"
#> 
#> $links[[41]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0236?format=json"
#> 
#> $links[[41]]$mediaType
#> [1] "application/json"
#> 
#> $links[[41]]$title
#> [1] "EO:EUM:DAT:0236"
#> 
#> $links[[41]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1A High Resolution (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[41]]$rel
#> [1] "collection/details"
#> 
#> $links[[41]]$numberOfProducts
#> [1] 12471
#> 
#> 
#> $links[[42]]
#> $links[[42]]$type
#> [1] "Link"
#> 
#> $links[[42]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0302?format=json"
#> 
#> $links[[42]]$mediaType
#> [1] "application/json"
#> 
#> $links[[42]]$title
#> [1] "EO:EUM:DAT:0302"
#> 
#> $links[[42]]$datasetTitle
#> [1] "GSA Level 2 Climate Data Record Release 2 - MFG - 63 degree"
#> 
#> $links[[42]]$rel
#> [1] "collection/details"
#> 
#> $links[[42]]$numberOfProducts
#> [1] 324
#> 
#> 
#> $links[[43]]
#> $links[[43]]$type
#> [1] "Link"
#> 
#> $links[[43]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0408?format=json"
#> 
#> $links[[43]]$mediaType
#> [1] "application/json"
#> 
#> $links[[43]]$title
#> [1] "EO:EUM:DAT:0408"
#> 
#> $links[[43]]$datasetTitle
#> [1] "OLCI Level 2 Ocean Colour Reduced Resolution - Sentinel-3"
#> 
#> $links[[43]]$rel
#> [1] "collection/details"
#> 
#> $links[[43]]$numberOfProducts
#> [1] 32278
#> 
#> 
#> $links[[44]]
#> $links[[44]]$type
#> [1] "Link"
#> 
#> $links[[44]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0409?format=json"
#> 
#> $links[[44]]$mediaType
#> [1] "application/json"
#> 
#> $links[[44]]$title
#> [1] "EO:EUM:DAT:0409"
#> 
#> $links[[44]]$datasetTitle
#> [1] "OLCI Level 1B Full Resolution - Sentinel-3"
#> 
#> $links[[44]]$rel
#> [1] "collection/details"
#> 
#> $links[[44]]$numberOfProducts
#> [1] 559571
#> 
#> 
#> $links[[45]]
#> $links[[45]]$type
#> [1] "Link"
#> 
#> $links[[45]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ACM%3AMETOP%3AASCSZOR02?format=json"
#> 
#> $links[[45]]$mediaType
#> [1] "application/json"
#> 
#> $links[[45]]$title
#> [1] "EO:EUM:CM:METOP:ASCSZOR02"
#> 
#> $links[[45]]$datasetTitle
#> [1] "ASCAT Level 1 SZO Climate Data Record Release 2 - Metop"
#> 
#> $links[[45]]$rel
#> [1] "collection/details"
#> 
#> $links[[45]]$numberOfProducts
#> [1] 37304
#> 
#> 
#> $links[[46]]
#> $links[[46]]$type
#> [1] "Link"
#> 
#> $links[[46]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ACM%3AMETOP%3AASCSZRR02?format=json"
#> 
#> $links[[46]]$mediaType
#> [1] "application/json"
#> 
#> $links[[46]]$title
#> [1] "EO:EUM:CM:METOP:ASCSZRR02"
#> 
#> $links[[46]]$datasetTitle
#> [1] "ASCAT Level 1 SZR Climate Data Record Release 2 - Metop"
#> 
#> $links[[46]]$rel
#> [1] "collection/details"
#> 
#> $links[[46]]$numberOfProducts
#> [1] 37304
#> 
#> 
#> $links[[47]]
#> $links[[47]]$type
#> [1] "Link"
#> 
#> $links[[47]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0558?format=json"
#> 
#> $links[[47]]$mediaType
#> [1] "application/json"
#> 
#> $links[[47]]$title
#> [1] "EO:EUM:DAT:0558"
#> 
#> $links[[47]]$datasetTitle
#> [1] "AVHRR GAC Atmospheric Motion Vectors Climate Data Record Release 2 - Multimission - Polar"
#> 
#> $links[[47]]$rel
#> [1] "collection/details"
#> 
#> $links[[47]]$numberOfProducts
#> [1] 42948
#> 
#> 
#> $links[[48]]
#> $links[[48]]$type
#> [1] "Link"
#> 
#> $links[[48]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0615?format=json"
#> 
#> $links[[48]]$mediaType
#> [1] "application/json"
#> 
#> $links[[48]]$title
#> [1] "EO:EUM:DAT:0615"
#> 
#> $links[[48]]$datasetTitle
#> [1] "SLSTR Level 1B Radiances and Brightness Temperatures (version BC004) - Sentinel-3 - Reprocessed"
#> 
#> $links[[48]]$rel
#> [1] "collection/details"
#> 
#> $links[[48]]$numberOfProducts
#> [1] 603404
#> 
#> 
#> $links[[49]]
#> $links[[49]]$type
#> [1] "Link"
#> 
#> $links[[49]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AOSI-150-A?format=json"
#> 
#> $links[[49]]$mediaType
#> [1] "application/json"
#> 
#> $links[[49]]$title
#> [1] "EO:EUM:DAT:METOP:OSI-150-A"
#> 
#> $links[[49]]$datasetTitle
#> [1] "ASCAT L2 25 km Winds Data Record Release 1 - Metop"
#> 
#> $links[[49]]$rel
#> [1] "collection/details"
#> 
#> $links[[49]]$numberOfProducts
#> [1] 37296
#> 
#> 
#> $links[[50]]
#> $links[[50]]$type
#> [1] "Link"
#> 
#> $links[[50]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AOSI-150-B?format=json"
#> 
#> $links[[50]]$mediaType
#> [1] "application/json"
#> 
#> $links[[50]]$title
#> [1] "EO:EUM:DAT:METOP:OSI-150-B"
#> 
#> $links[[50]]$datasetTitle
#> [1] "ASCAT L2 12.5 km Winds Data Record Release 1 - Metop"
#> 
#> $links[[50]]$rel
#> [1] "collection/details"
#> 
#> $links[[50]]$numberOfProducts
#> [1] 37302
#> 
#> 
#> $links[[51]]
#> $links[[51]]$type
#> [1] "Link"
#> 
#> $links[[51]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3ARSS-CLM?format=json"
#> 
#> $links[[51]]$mediaType
#> [1] "application/json"
#> 
#> $links[[51]]$title
#> [1] "EO:EUM:DAT:MSG:RSS-CLM"
#> 
#> $links[[51]]$datasetTitle
#> [1] "Rapid Scan Cloud Mask - MSG"
#> 
#> $links[[51]]$rel
#> [1] "collection/details"
#> 
#> $links[[51]]$numberOfProducts
#> [1] 1401069
#> 
#> 
#> $links[[52]]
#> $links[[52]]$type
#> [1] "Link"
#> 
#> $links[[52]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0834?format=json"
#> 
#> $links[[52]]$mediaType
#> [1] "application/json"
#> 
#> $links[[52]]$title
#> [1] "EO:EUM:DAT:0834"
#> 
#> $links[[52]]$datasetTitle
#> [1] "SRAL Level 2 Altimetry Global (version BC005) - Sentinel-3 - Reprocessed"
#> 
#> $links[[52]]$rel
#> [1] "collection/details"
#> 
#> $links[[52]]$numberOfProducts
#> [1] 121555
#> 
#> 
#> $links[[53]]
#> $links[[53]]$type
#> [1] "Link"
#> 
#> $links[[53]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0839?format=json"
#> 
#> $links[[53]]$mediaType
#> [1] "application/json"
#> 
#> $links[[53]]$title
#> [1] "EO:EUM:DAT:0839"
#> 
#> $links[[53]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B High Resolution (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[53]]$rel
#> [1] "collection/details"
#> 
#> $links[[53]]$numberOfProducts
#> [1] 19832
#> 
#> 
#> $links[[54]]
#> $links[[54]]$type
#> [1] "Link"
#> 
#> $links[[54]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0841?format=json"
#> 
#> $links[[54]]$mediaType
#> [1] "application/json"
#> 
#> $links[[54]]$title
#> [1] "EO:EUM:DAT:0841"
#> 
#> $links[[54]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 High Resolution (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[54]]$rel
#> [1] "collection/details"
#> 
#> $links[[54]]$numberOfProducts
#> [1] 19832
#> 
#> 
#> $links[[55]]
#> $links[[55]]$type
#> [1] "Link"
#> 
#> $links[[55]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0858?format=json"
#> 
#> $links[[55]]$mediaType
#> [1] "application/json"
#> 
#> $links[[55]]$title
#> [1] "EO:EUM:DAT:0858"
#> 
#> $links[[55]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2P High Resolution - Sentinel-6"
#> 
#> $links[[55]]$rel
#> [1] "collection/details"
#> 
#> $links[[55]]$numberOfProducts
#> [1] 51736
#> 
#> 
#> $links[[56]]
#> $links[[56]]$type
#> [1] "Link"
#> 
#> $links[[56]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0859?format=json"
#> 
#> $links[[56]]$mediaType
#> [1] "application/json"
#> 
#> $links[[56]]$title
#> [1] "EO:EUM:DAT:0859"
#> 
#> $links[[56]]$datasetTitle
#> [1] "Poseidon-4 Level 3 Altimetry High Resolution - Sentinel-6"
#> 
#> $links[[56]]$rel
#> [1] "collection/details"
#> 
#> $links[[56]]$numberOfProducts
#> [1] 1243
#> 
#> 
#> $links[[57]]
#> $links[[57]]$type
#> [1] "Link"
#> 
#> $links[[57]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0088?format=json"
#> 
#> $links[[57]]$mediaType
#> [1] "application/json"
#> 
#> $links[[57]]$title
#> [1] "EO:EUM:DAT:0088"
#> 
#> $links[[57]]$datasetTitle
#> [1] "Land Surface Temperature Climate Data Record - MSG"
#> 
#> $links[[57]]$rel
#> [1] "collection/details"
#> 
#> $links[[57]]$numberOfProducts
#> [1] 410197
#> 
#> 
#> $links[[58]]
#> $links[[58]]$type
#> [1] "Link"
#> 
#> $links[[58]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3AHRSEVIRI?format=json"
#> 
#> $links[[58]]$mediaType
#> [1] "application/json"
#> 
#> $links[[58]]$title
#> [1] "EO:EUM:DAT:MSG:HRSEVIRI"
#> 
#> $links[[58]]$datasetTitle
#> [1] "High Rate SEVIRI Level 1.5 Image Data - MSG - 0 degree"
#> 
#> $links[[58]]$rel
#> [1] "collection/details"
#> 
#> $links[[58]]$numberOfProducts
#> [1] 745935
#> 
#> 
#> $links[[59]]
#> $links[[59]]$type
#> [1] "Link"
#> 
#> $links[[59]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMSG%3AHRSEVIRI-IODC?format=json"
#> 
#> $links[[59]]$mediaType
#> [1] "application/json"
#> 
#> $links[[59]]$title
#> [1] "EO:EUM:DAT:MSG:HRSEVIRI-IODC"
#> 
#> $links[[59]]$datasetTitle
#> [1] "High Rate SEVIRI Level 1.5 Image Data - MSG - Indian Ocean"
#> 
#> $links[[59]]$rel
#> [1] "collection/details"
#> 
#> $links[[59]]$numberOfProducts
#> [1] 253084
#> 
#> 
#> $links[[60]]
#> $links[[60]]$type
#> [1] "Link"
#> 
#> $links[[60]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0300?format=json"
#> 
#> $links[[60]]$mediaType
#> [1] "application/json"
#> 
#> $links[[60]]$title
#> [1] "EO:EUM:DAT:0300"
#> 
#> $links[[60]]$datasetTitle
#> [1] "GSA Level 2 Climate Data Record Release 2 - MFG and MSG - 0 degree"
#> 
#> $links[[60]]$rel
#> [1] "collection/details"
#> 
#> $links[[60]]$numberOfProducts
#> [1] 1342
#> 
#> 
#> $links[[61]]
#> $links[[61]]$type
#> [1] "Link"
#> 
#> $links[[61]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0142?format=json"
#> 
#> $links[[61]]$mediaType
#> [1] "application/json"
#> 
#> $links[[61]]$title
#> [1] "EO:EUM:DAT:0142"
#> 
#> $links[[61]]$datasetTitle
#> [1] "Poseidon-4 Level 2P Wind/Wave Products Low Resolution in NRT - Sentinel-6"
#> 
#> $links[[61]]$rel
#> [1] "collection/details"
#> 
#> $links[[61]]$numberOfProducts
#> [1] 28721
#> 
#> 
#> $links[[62]]
#> $links[[62]]$type
#> [1] "Link"
#> 
#> $links[[62]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0238?format=json"
#> 
#> $links[[62]]$mediaType
#> [1] "application/json"
#> 
#> $links[[62]]$title
#> [1] "EO:EUM:DAT:0238"
#> 
#> $links[[62]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B High Resolution (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[62]]$rel
#> [1] "collection/details"
#> 
#> $links[[62]]$numberOfProducts
#> [1] 12471
#> 
#> 
#> $links[[63]]
#> $links[[63]]$type
#> [1] "Link"
#> 
#> $links[[63]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0374?format=json"
#> 
#> $links[[63]]$mediaType
#> [1] "application/json"
#> 
#> $links[[63]]$title
#> [1] "EO:EUM:DAT:0374"
#> 
#> $links[[63]]$datasetTitle
#> [1] "Commercial Radio Occultation Level 1B data - Lemur-2"
#> 
#> $links[[63]]$rel
#> [1] "collection/details"
#> 
#> $links[[63]]$numberOfProducts
#> [1] 41907
#> 
#> 
#> $links[[64]]
#> $links[[64]]$type
#> [1] "Link"
#> 
#> $links[[64]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0407?format=json"
#> 
#> $links[[64]]$mediaType
#> [1] "application/json"
#> 
#> $links[[64]]$title
#> [1] "EO:EUM:DAT:0407"
#> 
#> $links[[64]]$datasetTitle
#> [1] "OLCI Level 2 Ocean Colour Full Resolution - Sentinel-3"
#> 
#> $links[[64]]$rel
#> [1] "collection/details"
#> 
#> $links[[64]]$numberOfProducts
#> [1] 507660
#> 
#> 
#> $links[[65]]
#> $links[[65]]$type
#> [1] "Link"
#> 
#> $links[[65]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0411?format=json"
#> 
#> $links[[65]]$mediaType
#> [1] "application/json"
#> 
#> $links[[65]]$title
#> [1] "EO:EUM:DAT:0411"
#> 
#> $links[[65]]$datasetTitle
#> [1] "SLSTR Level 1B Radiances and Brightness Temperatures - Sentinel-3"
#> 
#> $links[[65]]$rel
#> [1] "collection/details"
#> 
#> $links[[65]]$numberOfProducts
#> [1] 1205568
#> 
#> 
#> $links[[66]]
#> $links[[66]]$type
#> [1] "Link"
#> 
#> $links[[66]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0417?format=json"
#> 
#> $links[[66]]$mediaType
#> [1] "application/json"
#> 
#> $links[[66]]$title
#> [1] "EO:EUM:DAT:0417"
#> 
#> $links[[66]]$datasetTitle
#> [1] "SLSTR Level 2 Fire Radiative Power - Sentinel 3"
#> 
#> $links[[66]]$rel
#> [1] "collection/details"
#> 
#> $links[[66]]$numberOfProducts
#> [1] 685535
#> 
#> 
#> $links[[67]]
#> $links[[67]]$type
#> [1] "Link"
#> 
#> $links[[67]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0601?format=json"
#> 
#> $links[[67]]$mediaType
#> [1] "application/json"
#> 
#> $links[[67]]$title
#> [1] "EO:EUM:DAT:0601"
#> 
#> $links[[67]]$datasetTitle
#> [1] "Poseidon-4 Level 3 Altimetry Low Resolution in NTC - Sentinel-6"
#> 
#> $links[[67]]$rel
#> [1] "collection/details"
#> 
#> $links[[67]]$numberOfProducts
#> [1] 376
#> 
#> 
#> $links[[68]]
#> $links[[68]]$type
#> [1] "Link"
#> 
#> $links[[68]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0556?format=json"
#> 
#> $links[[68]]$mediaType
#> [1] "application/json"
#> 
#> $links[[68]]$title
#> [1] "EO:EUM:DAT:0556"
#> 
#> $links[[68]]$datasetTitle
#> [1] "OLCI Level 2 Ocean Colour Full Resolution (version BC003) - Sentinel-3 - Reprocessed"
#> 
#> $links[[68]]$rel
#> [1] "collection/details"
#> 
#> $links[[68]]$numberOfProducts
#> [1] 707839
#> 
#> 
#> $links[[69]]
#> $links[[69]]$type
#> [1] "Link"
#> 
#> $links[[69]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0557?format=json"
#> 
#> $links[[69]]$mediaType
#> [1] "application/json"
#> 
#> $links[[69]]$title
#> [1] "EO:EUM:DAT:0557"
#> 
#> $links[[69]]$datasetTitle
#> [1] "OLCI Level 2 Ocean Colour Reduced Resolution (version BC003) - Sentinel-3 - Reprocessed"
#> 
#> $links[[69]]$rel
#> [1] "collection/details"
#> 
#> $links[[69]]$numberOfProducts
#> [1] 221187
#> 
#> 
#> $links[[70]]
#> $links[[70]]$type
#> [1] "Link"
#> 
#> $links[[70]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0576?format=json"
#> 
#> $links[[70]]$mediaType
#> [1] "application/json"
#> 
#> $links[[70]]$title
#> [1] "EO:EUM:DAT:0576"
#> 
#> $links[[70]]$datasetTitle
#> [1] "IASI All Sky Temperature and Humidity Profiles - Climate Data Record Release 1.1 - Metop-A and -B"
#> 
#> $links[[70]]$rel
#> [1] "collection/details"
#> 
#> $links[[70]]$numberOfProducts
#> [1] 122493
#> 
#> 
#> $links[[71]]
#> $links[[71]]$type
#> [1] "Link"
#> 
#> $links[[71]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0579?format=json"
#> 
#> $links[[71]]$mediaType
#> [1] "application/json"
#> 
#> $links[[71]]$title
#> [1] "EO:EUM:DAT:0579"
#> 
#> $links[[71]]$datasetTitle
#> [1] "Polar Multi-Sensor Aerosol Optical Properties Climate Data Record Release 1 - Metop-A and -B"
#> 
#> $links[[71]]$rel
#> [1] "collection/details"
#> 
#> $links[[71]]$numberOfProducts
#> [1] 88408
#> 
#> 
#> $links[[72]]
#> $links[[72]]$type
#> [1] "Link"
#> 
#> $links[[72]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0581?format=json"
#> 
#> $links[[72]]$mediaType
#> [1] "application/json"
#> 
#> $links[[72]]$title
#> [1] "EO:EUM:DAT:0581"
#> 
#> $links[[72]]$datasetTitle
#> [1] "SLSTR Level 1B Radiances and Brightness Temperatures (version BC003) - Sentinel-3 - Reprocessed"
#> 
#> $links[[72]]$rel
#> [1] "collection/details"
#> 
#> $links[[72]]$numberOfProducts
#> [1] 343732
#> 
#> 
#> $links[[73]]
#> $links[[73]]$type
#> [1] "Link"
#> 
#> $links[[73]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0583?format=json"
#> 
#> $links[[73]]$mediaType
#> [1] "application/json"
#> 
#> $links[[73]]$title
#> [1] "EO:EUM:DAT:0583"
#> 
#> $links[[73]]$datasetTitle
#> [1] "SRAL Level 1A Unpacked L0 Complex Echoes (version BC004) - Sentinel-3 - Reprocessed"
#> 
#> $links[[73]]$rel
#> [1] "collection/details"
#> 
#> $links[[73]]$numberOfProducts
#> [1] 54611
#> 
#> 
#> $links[[74]]
#> $links[[74]]$type
#> [1] "Link"
#> 
#> $links[[74]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0585?format=json"
#> 
#> $links[[74]]$mediaType
#> [1] "application/json"
#> 
#> $links[[74]]$title
#> [1] "EO:EUM:DAT:0585"
#> 
#> $links[[74]]$datasetTitle
#> [1] "SRAL Level 1B Stack Echoes (version BC004) - Sentinel-3 - Reprocessed"
#> 
#> $links[[74]]$rel
#> [1] "collection/details"
#> 
#> $links[[74]]$numberOfProducts
#> [1] 54610
#> 
#> 
#> $links[[75]]
#> $links[[75]]$type
#> [1] "Link"
#> 
#> $links[[75]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0853?format=json"
#> 
#> $links[[75]]$mediaType
#> [1] "application/json"
#> 
#> $links[[75]]$title
#> [1] "EO:EUM:DAT:0853"
#> 
#> $links[[75]]$datasetTitle
#> [1] "Radio Occultation Level 1B Products - Sentinel-6"
#> 
#> $links[[75]]$rel
#> [1] "collection/details"
#> 
#> $links[[75]]$numberOfProducts
#> [1] 852219
#> 
#> 
#> $links[[76]]
#> $links[[76]]$type
#> [1] "Link"
#> 
#> $links[[76]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0854?format=json"
#> 
#> $links[[76]]$mediaType
#> [1] "application/json"
#> 
#> $links[[76]]$title
#> [1] "EO:EUM:DAT:0854"
#> 
#> $links[[76]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 Low Resolution - Sentinel-6"
#> 
#> $links[[76]]$rel
#> [1] "collection/details"
#> 
#> $links[[76]]$numberOfProducts
#> [1] 11357
#> 
#> 
#> $links[[77]]
#> $links[[77]]$type
#> [1] "Link"
#> 
#> $links[[77]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0837?format=json"
#> 
#> $links[[77]]$mediaType
#> [1] "application/json"
#> 
#> $links[[77]]$title
#> [1] "EO:EUM:DAT:0837"
#> 
#> $links[[77]]$datasetTitle
#> [1] "Climate-quality Advanced Microwave Radiometer Level 2 Products (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[77]]$rel
#> [1] "collection/details"
#> 
#> $links[[77]]$numberOfProducts
#> [1] 20060
#> 
#> 
#> $links[[78]]
#> $links[[78]]$type
#> [1] "Link"
#> 
#> $links[[78]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0838?format=json"
#> 
#> $links[[78]]$mediaType
#> [1] "application/json"
#> 
#> $links[[78]]$title
#> [1] "EO:EUM:DAT:0838"
#> 
#> $links[[78]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1A High Resolution (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[78]]$rel
#> [1] "collection/details"
#> 
#> $links[[78]]$numberOfProducts
#> [1] 19835
#> 
#> 
#> $links[[79]]
#> $links[[79]]$type
#> [1] "Link"
#> 
#> $links[[79]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0840?format=json"
#> 
#> $links[[79]]$mediaType
#> [1] "application/json"
#> 
#> $links[[79]]$title
#> [1] "EO:EUM:DAT:0840"
#> 
#> $links[[79]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B Low Resolution (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[79]]$rel
#> [1] "collection/details"
#> 
#> $links[[79]]$numberOfProducts
#> [1] 20000
#> 
#> 
#> $links[[80]]
#> $links[[80]]$type
#> [1] "Link"
#> 
#> $links[[80]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AMHSL1?format=json"
#> 
#> $links[[80]]$mediaType
#> [1] "application/json"
#> 
#> $links[[80]]$title
#> [1] "EO:EUM:DAT:METOP:MHSL1"
#> 
#> $links[[80]]$datasetTitle
#> [1] "MHS Level 1B - Metop - Global"
#> 
#> $links[[80]]$rel
#> [1] "collection/details"
#> 
#> $links[[80]]$numberOfProducts
#> [1] 160391
#> 
#> 
#> $links[[81]]
#> $links[[81]]$type
#> [1] "Link"
#> 
#> $links[[81]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AIASIL1C-ALL?format=json"
#> 
#> $links[[81]]$mediaType
#> [1] "application/json"
#> 
#> $links[[81]]$title
#> [1] "EO:EUM:DAT:METOP:IASIL1C-ALL"
#> 
#> $links[[81]]$datasetTitle
#> [1] "IASI Level 1C - All Spectral Samples - Metop - Global"
#> 
#> $links[[81]]$rel
#> [1] "collection/details"
#> 
#> $links[[81]]$numberOfProducts
#> [1] 96296
#> 
#> 
#> $links[[82]]
#> $links[[82]]$type
#> [1] "Link"
#> 
#> $links[[82]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AIASSND02?format=json"
#> 
#> $links[[82]]$mediaType
#> [1] "application/json"
#> 
#> $links[[82]]$title
#> [1] "EO:EUM:DAT:METOP:IASSND02"
#> 
#> $links[[82]]$datasetTitle
#> [1] "IASI Combined Sounding Products - Metop"
#> 
#> $links[[82]]$rel
#> [1] "collection/details"
#> 
#> $links[[82]]$numberOfProducts
#> [1] 149405
#> 
#> 
#> $links[[83]]
#> $links[[83]]$type
#> [1] "Link"
#> 
#> $links[[83]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0758?format=json"
#> 
#> $links[[83]]$mediaType
#> [1] "application/json"
#> 
#> $links[[83]]$title
#> [1] "EO:EUM:DAT:0758"
#> 
#> $links[[83]]$datasetTitle
#> [1] "IASI Principal Components Scores Fundamental Data Record Release 1 - Metop-A and -B"
#> 
#> $links[[83]]$rel
#> [1] "collection/details"
#> 
#> $links[[83]]$numberOfProducts
#> [1] 122636
#> 
#> 
#> $links[[84]]
#> $links[[84]]$type
#> [1] "Link"
#> 
#> $links[[84]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0881?format=json"
#> 
#> $links[[84]]$mediaType
#> [1] "application/json"
#> 
#> $links[[84]]$title
#> [1] "EO:EUM:DAT:0881"
#> 
#> $links[[84]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 2 - MFG - 57 degree"
#> 
#> $links[[84]]$rel
#> [1] "collection/details"
#> 
#> $links[[84]]$numberOfProducts
#> [1] 176726
#> 
#> 
#> $links[[85]]
#> $links[[85]]$type
#> [1] "Link"
#> 
#> $links[[85]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0080?format=json"
#> 
#> $links[[85]]$mediaType
#> [1] "application/json"
#> 
#> $links[[85]]$title
#> [1] "EO:EUM:DAT:0080"
#> 
#> $links[[85]]$datasetTitle
#> [1] "MVIRI Level 1.5 Climate Data Record Release 1 - MFG - 0 degree"
#> 
#> $links[[85]]$rel
#> [1] "collection/details"
#> 
#> $links[[85]]$numberOfProducts
#> [1] 427873
#> 
#> 
#> $links[[86]]
#> $links[[86]]$type
#> [1] "Link"
#> 
#> $links[[86]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0240?format=json"
#> 
#> $links[[86]]$mediaType
#> [1] "application/json"
#> 
#> $links[[86]]$title
#> [1] "EO:EUM:DAT:0240"
#> 
#> $links[[86]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 High Resolution (baseline version F06) - Sentinel-6 - Reprocessed"
#> 
#> $links[[86]]$rel
#> [1] "collection/details"
#> 
#> $links[[86]]$numberOfProducts
#> [1] 12471
#> 
#> 
#> $links[[87]]
#> $links[[87]]$type
#> [1] "Link"
#> 
#> $links[[87]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0533?format=json"
#> 
#> $links[[87]]$mediaType
#> [1] "application/json"
#> 
#> $links[[87]]$title
#> [1] "EO:EUM:DAT:0533"
#> 
#> $links[[87]]$datasetTitle
#> [1] "GOME-2 Level 1B Fundamental Data Record Release 3 - Metop-A and -B"
#> 
#> $links[[87]]$rel
#> [1] "collection/details"
#> 
#> $links[[87]]$numberOfProducts
#> [1] 101404
#> 
#> 
#> $links[[88]]
#> $links[[88]]$type
#> [1] "Link"
#> 
#> $links[[88]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0577?format=json"
#> 
#> $links[[88]]$mediaType
#> [1] "application/json"
#> 
#> $links[[88]]$title
#> [1] "EO:EUM:DAT:0577"
#> 
#> $links[[88]]$datasetTitle
#> [1] "OLCI Level 1B Full Resolution (version BC002) - Sentinel-3 - Reprocessed"
#> 
#> $links[[88]]$rel
#> [1] "collection/details"
#> 
#> $links[[88]]$numberOfProducts
#> [1] 307062
#> 
#> 
#> $links[[89]]
#> $links[[89]]$type
#> [1] "Link"
#> 
#> $links[[89]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0578?format=json"
#> 
#> $links[[89]]$mediaType
#> [1] "application/json"
#> 
#> $links[[89]]$title
#> [1] "EO:EUM:DAT:0578"
#> 
#> $links[[89]]$datasetTitle
#> [1] "OLCI Level 1B Reduced Resolution (version BC002) - Sentinel-3 - Reprocessed"
#> 
#> $links[[89]]$rel
#> [1] "collection/details"
#> 
#> $links[[89]]$numberOfProducts
#> [1] 195825
#> 
#> 
#> $links[[90]]
#> $links[[90]]$type
#> [1] "Link"
#> 
#> $links[[90]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0582?format=json"
#> 
#> $links[[90]]$mediaType
#> [1] "application/json"
#> 
#> $links[[90]]$title
#> [1] "EO:EUM:DAT:0582"
#> 
#> $links[[90]]$datasetTitle
#> [1] "SLSTR Level 2 Sea Surface Temperature (SST) (version BC003) - Sentinel-3 - Reprocessed"
#> 
#> $links[[90]]$rel
#> [1] "collection/details"
#> 
#> $links[[90]]$numberOfProducts
#> [1] 9968
#> 
#> 
#> $links[[91]]
#> $links[[91]]$type
#> [1] "Link"
#> 
#> $links[[91]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0586?format=json"
#> 
#> $links[[91]]$mediaType
#> [1] "application/json"
#> 
#> $links[[91]]$title
#> [1] "EO:EUM:DAT:0586"
#> 
#> $links[[91]]$datasetTitle
#> [1] "SRAL Level 2 Altimetry Global (version BC004) - Sentinel-3 - Reprocessed"
#> 
#> $links[[91]]$rel
#> [1] "collection/details"
#> 
#> $links[[91]]$numberOfProducts
#> [1] 57053
#> 
#> 
#> $links[[92]]
#> $links[[92]]$type
#> [1] "Link"
#> 
#> $links[[92]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0414?format=json"
#> 
#> $links[[92]]$mediaType
#> [1] "application/json"
#> 
#> $links[[92]]$title
#> [1] "EO:EUM:DAT:0414"
#> 
#> $links[[92]]$datasetTitle
#> [1] "SRAL Level 1B stack echoes - Sentinel-3"
#> 
#> $links[[92]]$rel
#> [1] "collection/details"
#> 
#> $links[[92]]$numberOfProducts
#> [1] 24170
#> 
#> 
#> $links[[93]]
#> $links[[93]]$type
#> [1] "Link"
#> 
#> $links[[93]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0842?format=json"
#> 
#> $links[[93]]$mediaType
#> [1] "application/json"
#> 
#> $links[[93]]$title
#> [1] "EO:EUM:DAT:0842"
#> 
#> $links[[93]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 2 Low Resolution (baseline version F08) - Sentinel-6 - Reprocessed"
#> 
#> $links[[93]]$rel
#> [1] "collection/details"
#> 
#> $links[[93]]$numberOfProducts
#> [1] 19996
#> 
#> 
#> $links[[94]]
#> $links[[94]]$type
#> [1] "Link"
#> 
#> $links[[94]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0850?format=json"
#> 
#> $links[[94]]$mediaType
#> [1] "application/json"
#> 
#> $links[[94]]$title
#> [1] "EO:EUM:DAT:0850"
#> 
#> $links[[94]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1A High Resolution - Sentinel-6"
#> 
#> $links[[94]]$rel
#> [1] "collection/details"
#> 
#> $links[[94]]$numberOfProducts
#> [1] 15508
#> 
#> 
#> $links[[95]]
#> $links[[95]]$type
#> [1] "Link"
#> 
#> $links[[95]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0852?format=json"
#> 
#> $links[[95]]$mediaType
#> [1] "application/json"
#> 
#> $links[[95]]$title
#> [1] "EO:EUM:DAT:0852"
#> 
#> $links[[95]]$datasetTitle
#> [1] "Poseidon-4 Altimetry Level 1B Low Resolution - Sentinel-6"
#> 
#> $links[[95]]$rel
#> [1] "collection/details"
#> 
#> $links[[95]]$numberOfProducts
#> [1] 11359
#> 
#> 
#> $links[[96]]
#> $links[[96]]$type
#> [1] "Link"
#> 
#> $links[[96]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AOSI-104?format=json"
#> 
#> $links[[96]]$mediaType
#> [1] "application/json"
#> 
#> $links[[96]]$title
#> [1] "EO:EUM:DAT:METOP:OSI-104"
#> 
#> $links[[96]]$datasetTitle
#> [1] "ASCAT Coastal Winds at 12.5 km Swath Grid - Metop"
#> 
#> $links[[96]]$rel
#> [1] "collection/details"
#> 
#> $links[[96]]$numberOfProducts
#> [1] 115472
#> 
#> 
#> $links[[97]]
#> $links[[97]]$type
#> [1] "Link"
#> 
#> $links[[97]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3ASOMO12?format=json"
#> 
#> $links[[97]]$mediaType
#> [1] "application/json"
#> 
#> $links[[97]]$title
#> [1] "EO:EUM:DAT:METOP:SOMO12"
#> 
#> $links[[97]]$datasetTitle
#> [1] "ASCAT Soil Moisture at 12.5 km Swath Grid in NRT - Metop"
#> 
#> $links[[97]]$rel
#> [1] "collection/details"
#> 
#> $links[[97]]$numberOfProducts
#> [1] 154632
#> 
#> 
#> $links[[98]]
#> $links[[98]]$type
#> [1] "Link"
#> 
#> $links[[98]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3ASOMO25?format=json"
#> 
#> $links[[98]]$mediaType
#> [1] "application/json"
#> 
#> $links[[98]]$title
#> [1] "EO:EUM:DAT:METOP:SOMO25"
#> 
#> $links[[98]]$datasetTitle
#> [1] "ASCAT Soil Moisture at 25 km Swath Grid in NRT - Metop"
#> 
#> $links[[98]]$rel
#> [1] "collection/details"
#> 
#> $links[[98]]$numberOfProducts
#> [1] 154632
#> 
#> 
#> $links[[99]]
#> $links[[99]]$type
#> [1] "Link"
#> 
#> $links[[99]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AAMSUL1?format=json"
#> 
#> $links[[99]]$mediaType
#> [1] "application/json"
#> 
#> $links[[99]]$title
#> [1] "EO:EUM:DAT:METOP:AMSUL1"
#> 
#> $links[[99]]$datasetTitle
#> [1] "AMSU-A Level 1B - Metop - Global"
#> 
#> $links[[99]]$rel
#> [1] "collection/details"
#> 
#> $links[[99]]$numberOfProducts
#> [1] 161312
#> 
#> 
#> $links[[100]]
#> $links[[100]]$type
#> [1] "Link"
#> 
#> $links[[100]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3AAVHRRL1?format=json"
#> 
#> $links[[100]]$mediaType
#> [1] "application/json"
#> 
#> $links[[100]]$title
#> [1] "EO:EUM:DAT:METOP:AVHRRL1"
#> 
#> $links[[100]]$datasetTitle
#> [1] "AVHRR Level 1B - Metop - Global"
#> 
#> $links[[100]]$rel
#> [1] "collection/details"
#> 
#> $links[[100]]$numberOfProducts
#> [1] 161334
#> 
#> 
#> $links[[101]]
#> $links[[101]]$type
#> [1] "Link"
#> 
#> $links[[101]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3AMETOP%3ALSA-002?format=json"
#> 
#> $links[[101]]$mediaType
#> [1] "application/json"
#> 
#> $links[[101]]$title
#> [1] "EO:EUM:DAT:METOP:LSA-002"
#> 
#> $links[[101]]$datasetTitle
#> [1] "Daily Land Surface Temperature - Metop"
#> 
#> $links[[101]]$rel
#> [1] "collection/details"
#> 
#> $links[[101]]$numberOfProducts
#> [1] 525
#> 
#> 
#> $links[[102]]
#> $links[[102]]$type
#> [1] "Link"
#> 
#> $links[[102]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3A0645?format=json"
#> 
#> $links[[102]]$mediaType
#> [1] "application/json"
#> 
#> $links[[102]]$title
#> [1] "EO:EUM:DAT:0645"
#> 
#> $links[[102]]$datasetTitle
#> [1] "Global Sea Ice Concentration Interim Climate Data Record Release 3 - DMSP"
#> 
#> $links[[102]]$rel
#> [1] "collection/details"
#> 
#> $links[[102]]$numberOfProducts
#> [1] 401
#> 
#> 
#> $links[[103]]
#> $links[[103]]$type
#> [1] "Link"
#> 
#> $links[[103]]$href
#> [1] "https://api.eumetsat.int/data/browse/1.0.0/collections/EO%3AEUM%3ADAT%3ADMSP%3AOSI-401-B?format=json"
#> 
#> $links[[103]]$mediaType
#> [1] "application/json"
#> 
#> $links[[103]]$title
#> [1] "EO:EUM:DAT:DMSP:OSI-401-B"
#> 
#> $links[[103]]$datasetTitle
#> [1] "Global Sea Ice Concentration (netCDF) - DMSP"
#> 
#> $links[[103]]$rel
#> [1] "collection/details"
#> 
#> $links[[103]]$numberOfProducts
#> [1] 68
#> 
#> 
#> 
#> $numberOfProducts
#> [1] 28652435
#>